Started 1.20 port

+++ src/main/java/com/ninni/species/client/particles/BirtdParticle.java
@@ -47,7 +47,7 @@ public class BirtdParticle extends TextureSheetParticle {
 
     @Override
     protected int getLightColor(float f) {
-        BlockPos blockPos = new BlockPos(this.x, this.y, this.z);
+        BlockPos blockPos = new BlockPos((int)this.x, (int)this.y, (int)this.z);
         if (this.level.hasChunkAt(blockPos)) {
             return LevelRenderer.getLightColor(this.level, blockPos);
         }
